Refactor demo & demoUtils
