+++ src/demo/demoUtils.js
@@ -214,3 +214,38 @@ export const commonLanguages = {
 }
 
 export const commonRegions = ['EU', 'US', 'CA']
+
+export const getTokenFactoryUrl = (region) => {
+  switch (region) {
+    case 'US':
+      return process.env.US_JWT_FACTORY
+    case 'CA':
+      return process.env.CA_JWT_FACTORY
+    default:
+      return process.env.JWT_FACTORY
+  }
+}
+
+export const getToken = (hasPreview, url, eventEmitter, onSuccess) => {
+  const request = new XMLHttpRequest()
+  request.open('GET', url, true)
+  request.setRequestHeader(
+    'Authorization',
+    `BASIC ${process.env.SDK_TOKEN_FACTORY_SECRET}`
+  )
+  request.onload = function () {
+    if (request.status >= 200 && request.status < 400) {
+      const data = JSON.parse(request.responseText)
+      if (hasPreview && eventEmitter) {
+        eventEmitter.postMessage({
+          type: 'UPDATE_CHECK_DATA',
+          payload: {
+            applicantId: data.applicant_id,
+          },
+        })
+      }
+      onSuccess(data.message)
+    }
+  }
+  request.send()
+}
